Cleanup neg syncers:
- split l4//l7 neg endpoint calculators into separate files
- refactor l4 neg calculator
